worker/requirements.txt

psycopg2-binary==2.9.2
pydantic==1.8.2